BentoML/requirements.txt
transformers
bentoml
torch
pillow
yaml
requests